\name{TERGMdir-package}
\alias{TERGMdir-package}
\alias{TERGMdir}
\docType{package}
\title{
\packageTitle{TERGMdir}
}
\description{
\packageDescription{TERGMdir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TERGMdir}
\packageIndices{TERGMdir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TERGMdir}

Maintainer: \packageMaintainer{TERGMdir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
